Completely change project structure
