+++ Sources/Abstract/Morphism/Isomorphism.swift
