+++ Sources/Abstract/Protocols/Magma.swift
