+++ Sources/Abstract/Protocols/Semigroup.swift
@@ -1,0 +1,25 @@
+/*:
+# Semigroup
+
+A Semigroup is a Magma where the composition operation is associative, which means that multiple operations in sequence can happen in any time order.
+
+To put it simply: (a <> b) <> c = a <> (b <> c)
+*/
+
+#if SWIFT_PACKAGE
+    import Operadics
+#endif
+
+public protocol Semigroup: Magma {}
+
+extension Law where Element: Semigroup {
+	public static func isAssociative(_ a: Element, _ b: Element, _ c: Element) -> Bool {
+		return (a <> b <> c) == (a <> (b <> c))
+	}
+}
+
+extension LawInContext where Element: Semigroup {
+	public static func isAssociative(_ a: Element, _ b: Element, _ c: Element) -> (Element.Context) -> Bool {
+		return (a <> b <> c) == (a <> (b <> c))
+	}
+}
