+++ Sources/Abstract/Support/Adapters.swift
