+++ Sources/Abstract/Support/Law.swift
