+++ Sources/Abstract/Types/String.swift
@@ -1,0 +1,11 @@
+#if SWIFT_PACKAGE
+	import Operadics
+#endif
+
+extension String: Monoid {
+	public static let empty: String = ""
+
+	public static func <> (left: String, right: String) -> String {
+		return left + right
+	}
+}
